+++ Source/SwiftLintFramework/Rules/Idiomatic/DuplicateImportsRuleExamples.swift
@@ -2,6 +2,8 @@ internal struct DuplicateImportsRuleExamples {
     static let nonTriggeringExamples: [Example] = [
         Example("import A\nimport B\nimport C"),
         Example("import A.B\nimport A.C"),
+        Example("@_implementationOnly import A\n@_implementationOnly import B"),
+        Example("@testable import A\n@testable import B"),
         Example("""
         #if DEBUG
             @testable import KsApi
@@ -21,6 +23,8 @@ internal struct DuplicateImportsRuleExamples {
         var list: [Example] = [
             Example("import Foundation\nimport Dispatch\n↓import Foundation"),
             Example("import Foundation\n↓import Foundation.NSString"),
+            Example("@_implementationOnly import A\n@_implementationOnly import A"),
+            Example("@testable import A\n@testable import A"),
             Example("↓import Foundation.NSString\nimport Foundation"),
             Example("↓import A.B.C\nimport A.B"),
             Example("import A.B\n↓import A.B.C"),