+++ CHANGELOG.md
@@ -41,10 +41,14 @@
 * Rule `unused_capture_list` should not be triggered by unowned self keyword.  
   [hank121314](https://github.com/hank121314)
   [#3389](https://github.com/realm/SwiftLint/issues/3389)
-  
+
 * Fix severity configuration for `indentation_width`.  
   [Samasaur1](https://github.com/Samasaur1), issue 
   [#3346](https://github.com/realm/SwiftLint/issues/3389)
+
+* Fix DuplicateImportsRule's support for import attributes.  
+  [Keith Smiley](https://github.com/keith)
+  [#3402](https://github.com/realm/SwiftLint/issues/3402)
 
 ## 0.40.3: Greased Up Drum Bearings